Refactor functions positions.
